clean: one entry point. python server.py. done.
Deleted: boot.py, elastik.py, lucy.py, 8 scripts.
Merged: boot.py startup logic → server.py __main__ (6 lines).
-1559 lines. No more "which file do I run?"
