+++ .github/workflows/plugins.yml
@@ -7,7 +7,6 @@ on:
       - 'plugins/**'
       - 'go/native/**'
       - 'server.py'
-      - 'boot.py'
       - 'plugins.py'
       - 'tests/test_plugins.py'
       - '.github/workflows/plugins.yml'
@@ -16,7 +15,6 @@ on:
       - 'plugins/**'
       - 'go/native/**'
       - 'server.py'
-      - 'boot.py'
       - 'plugins.py'
       - 'tests/test_plugins.py'
       - '.github/workflows/plugins.yml'
