+++ Dockerfile.dev
@@ -5,4 +5,4 @@ COPY requirements.txt .
 RUN pip install -r requirements.txt
 EXPOSE 3004
 ENV PYTHONUNBUFFERED=1
-CMD ["python", "boot.py"]
+CMD ["python", "server.py"]
